Solve day 10 part 1
Part 2 is going to be difficult. I'm currently thinking that I'd run a
flood fill from the outside edge, and then get

enclosed = total - outside - pipe loop

The problem here is that a flood fill requires some data storage for
the visited set, so that's memory.

I also need to find the boundaries of the pipe loop, because
re-tracing the loop on every flood fill step is not going to be
viable. My loop is around 13.5k steps long. If I wanted to keep that
in memory I could only do it if every coordinate pair is two bytes or
less. With XY pairs that would be (u8, u8), or otherwise u16 as
absolute indexing. Both of which are technically good enough to cover
the 140x140 total area. But even then, we're talking about 27 kB of
memory just for the pipe loop, which might not leave enough for the
flood fill and remaining stack.

And because this is no alloc, I have to statically commit to all these
sizes.