Speed up day 14 by some 90%
Generating the grouping map in a loop is really no good, and even doing
it manually using FxHashMap isn't, it's actually even slower. Just
counting Y-coordinate popularity via an array is much faster, and then
just looking at relevant robots with that Y-coordinate, but only if that
row is even populated enough.