Add day 16
This is a horrible solution, but it's correct and runs in finite time
(~20 seconds). I might explore a recursive solution with a visited
vector instead, to avoid having to copy the visited hash set over and
over for part 2.

Part 1 actually used a single shared visited hash set, because it just
needed to find any one of the optimal paths, but part 2 needs per-path
tracking so that two optimal paths can converge. Or maybe I can just do
this through the cost tracking?